split bc into files
