+++ demos/FiniteVolume/manual_block_matrix_assembly.cpp
@@ -4,7 +4,6 @@
 #include <samurai/io/hdf5.hpp>
 #include <samurai/mr/adapt.hpp>
 #include <samurai/mr/mesh.hpp>
-#include <samurai/samurai.hpp>
 #include <samurai/schemes/fv.hpp>
 
 using aux_t = xt::xtensor<double, 2>;
